+++ history.md
@@ -1,7 +1,9 @@
 # 2022-05-07
 
 - Show sort method as description in outline view.
-- Add configuration `outline.switchSortKey`, default to `<C-s>`
+- Add configuration `outline.switchSortKey`, default to `<C-s>`.
+- Add configuration `outline.detailAsDescription`, default to `true`.
+- Add variable `g:coc_max_treeview_width`.
 
 # 2022-05-06